refactor service task behavior to its own class
