+++ jbpm/jbpm-flow-builder/src/main/java/org/jbpm/compiler/canonical/WorkItemNodeVisitor.java
@@ -99,7 +99,16 @@ public class WorkItemNodeVisitor<T extends WorkItemNode> extends AbstractNodeVis
     @Override
     public void visitNode(String factoryField, T node, BlockStmt body, VariableScope variableScope, ProcessMetaData metadata) {
         Work work = node.getWork();
-        String workName = workItemName(node, metadata);
+        String workName = node.getWork().getName();
+
+        if (workName.equals("Service Task")) {
+            ServiceTaskDescriptor d = new ServiceTaskDescriptor(node, contextClassLoader);
+            String mangledName = d.mangledName();
+            CompilationUnit generatedHandler = d.generateHandlerClassForService();
+            metadata.getGeneratedHandlers().put(mangledName, generatedHandler);
+            workName = mangledName;
+        }
+
         body.addStatement(getAssignedFactoryMethod(factoryField, WorkItemNodeFactory.class, getNodeId(node), getNodeKey(), new LongLiteralExpr(node.getId())))
                 .addStatement(getNameMethod(node, work.getName()))
                 .addStatement(getFactoryMethod(getNodeId(node), METHOD_WORK_NAME, new StringLiteralExpr(workName)));
@@ -147,155 +156,6 @@ public class WorkItemNodeVisitor<T extends WorkItemNode> extends AbstractNodeVis
         }
     }
 
-    protected String workItemName(WorkItemNode workItemNode, ProcessMetaData metadata) {
-        String workName = workItemNode.getWork().getName();
 
-        if (workName.equals("Service Task")) {
-            String interfaceName = (String) workItemNode.getWork().getParameter("Interface");
-            String operationName = (String) workItemNode.getWork().getParameter("Operation");
-            String type = (String) workItemNode.getWork().getParameter("ParameterType");
 
-            NodeValidator.of(getNodeKey(), workItemNode.getName())
-                    .notEmpty("interfaceName", interfaceName)
-                    .notEmpty("operationName", operationName)
-                    .validate();
-
-            Map<String, String> parameters = null;
-            if (type != null) {
-                if (isDefaultParameterType(type)) {
-                    type = inferParameterType(workItemNode.getName(), interfaceName, operationName, type);
-                }
-
-                parameters = Collections.singletonMap("Parameter", type);
-            } else {
-                parameters = new LinkedHashMap<>();
-
-                for (ParameterDefinition def : workItemNode.getWork().getParameterDefinitions()) {
-                    parameters.put(def.getName(), def.getType().getStringType());
-                }
-            }
-
-            String uniqueHandlerName = mangledHandlerName(interfaceName, operationName, parameters);
-
-            CompilationUnit handlerClass = generateHandlerClassForService(
-                    uniqueHandlerName, interfaceName, operationName, parameters, workItemNode.getOutAssociations());
-
-            metadata.getGeneratedHandlers().put(uniqueHandlerName, handlerClass);
-
-            return uniqueHandlerName;
-        }
-
-        return workName;
-    }
-
-    private String mangledHandlerName(String interfaceName, String operationName, Map<String, String> parameters) {
-        String simpleName = interfaceName.substring(interfaceName.lastIndexOf(".") + 1);
-
-        // mangle dotted identifiers foo.bar.Baz into foo$bar$Baz
-        // then concatenate the collection with $$
-        // e.g. List.of("foo.bar.Baz", "qux.Quux") -> "foo$bar$Baz$$qux$Quux"
-        String mangledParameterTypes =
-                parameters.values().stream().map(s -> s.replace('.', '$'))
-                        .collect(joining("$$"));
-
-        return String.format("%s_%s_%s_Handler", simpleName, operationName, mangledParameterTypes);
-    }
-
-    // assume 1 single arg as above
-    private String inferParameterType(String nodeName, String interfaceName, String operationName, String defaultType) {
-        try {
-            Class<?> i = contextClassLoader.loadClass(interfaceName);
-            for (Method m : i.getMethods()) {
-                if (m.getName().equals(operationName) && m.getParameterCount() == 1) {
-                    return m.getParameterTypes()[0].getCanonicalName();
-                }
-            }
-        } catch (ClassNotFoundException e) {
-            throw new IllegalArgumentException(MessageFormat.format("Invalid work item \"{0}\": class not found for interfaceName \"{1}\"", nodeName, interfaceName));
-        }
-        throw new IllegalArgumentException(MessageFormat.format("Invalid work item \"{0}\": could not find a method called \"{1}\" in class \"{2}\"", nodeName, operationName, interfaceName));
-    }
-
-    private boolean isDefaultParameterType(String type) {
-        return type.equals("java.lang.Object") || type.equals("Object");
-    }
-
-    protected CompilationUnit generateHandlerClassForService(String mangledHandlerName, String interfaceName, String operation, Map<String, String> parameters, List<DataAssociation> outAssociations) {
-        CompilationUnit compilationUnit = new CompilationUnit("org.kie.kogito.handlers");
-
-        compilationUnit.getTypes().add(classDeclaration(mangledHandlerName, interfaceName, operation, parameters, outAssociations));
-
-        return compilationUnit;
-    }
-
-    public ClassOrInterfaceDeclaration classDeclaration(String mangledHandlerName, String interfaceName, String operation, Map<String, String> parameters, List<DataAssociation> outAssociations) {
-        ClassOrInterfaceDeclaration cls = new ClassOrInterfaceDeclaration()
-                .setName(mangledHandlerName)
-                .setModifiers(Modifier.Keyword.PUBLIC)
-                .addImplementedType(WorkItemHandler.class.getCanonicalName());
-        ClassOrInterfaceType serviceType = new ClassOrInterfaceType(null, interfaceName);
-        FieldDeclaration serviceField = new FieldDeclaration()
-                .addVariable(new VariableDeclarator(serviceType, "service"));
-        cls.addMember(serviceField);
-
-        // executeWorkItem method
-        BlockStmt executeWorkItemBody = new BlockStmt();
-        MethodDeclaration executeWorkItem = new MethodDeclaration()
-                .setModifiers(Modifier.Keyword.PUBLIC)
-                .setType(void.class)
-                .setName("executeWorkItem")
-                .setBody(executeWorkItemBody)
-                .addParameter(WorkItem.class.getCanonicalName(), "workItem")
-                .addParameter(WorkItemManager.class.getCanonicalName(), "workItemManager");
-
-        MethodCallExpr callService = new MethodCallExpr(new NameExpr("service"), operation);
-
-        for (Entry<String, String> paramEntry : parameters.entrySet()) {
-            MethodCallExpr getParamMethod = new MethodCallExpr(new NameExpr("workItem"), "getParameter").addArgument(new StringLiteralExpr(paramEntry.getKey()));
-            callService.addArgument(new CastExpr(new ClassOrInterfaceType(null, paramEntry.getValue()), getParamMethod));
-        }
-        Expression results = null;
-        if (outAssociations.isEmpty()) {
-
-            executeWorkItemBody.addStatement(callService);
-            results = new NullLiteralExpr();
-        } else {
-            VariableDeclarationExpr resultField = new VariableDeclarationExpr()
-                    .addVariable(new VariableDeclarator(new ClassOrInterfaceType(null, Object.class.getCanonicalName()), "result", callService));
-
-            executeWorkItemBody.addStatement(resultField);
-
-            results = new MethodCallExpr(new NameExpr("java.util.Collections"), "singletonMap")
-                    .addArgument(new StringLiteralExpr(outAssociations.get(0).getSources().get(0)))
-                    .addArgument(new NameExpr("result"));
-        }
-
-        MethodCallExpr completeWorkItem = new MethodCallExpr(new NameExpr("workItemManager"), "completeWorkItem")
-                .addArgument(new MethodCallExpr(new NameExpr("workItem"), "getId"))
-                .addArgument(results);
-
-        executeWorkItemBody.addStatement(completeWorkItem);
-
-        // abortWorkItem method
-        BlockStmt abortWorkItemBody = new BlockStmt();
-        MethodDeclaration abortWorkItem = new MethodDeclaration()
-                .setModifiers(Modifier.Keyword.PUBLIC)
-                .setType(void.class)
-                .setName("abortWorkItem")
-                .setBody(abortWorkItemBody)
-                .addParameter(WorkItem.class.getCanonicalName(), "workItem")
-                .addParameter(WorkItemManager.class.getCanonicalName(), "workItemManager");
-
-        // getName method
-        MethodDeclaration getName = new MethodDeclaration()
-                .setModifiers(Modifier.Keyword.PUBLIC)
-                .setType(String.class)
-                .setName("getName")
-                .setBody(new BlockStmt().addStatement(new ReturnStmt(new StringLiteralExpr(interfaceName + "." + operation))));
-        cls.addMember(executeWorkItem)
-                .addMember(abortWorkItem)
-                .addMember(getName);
-
-        return cls;
-    }
 }
